hackathon/hackathon/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-genai==1.16.1
requests==2.31.0 